+++ lib/core/internals/index.ts
@@ -1,32 +1,38 @@
 
-import { CoreOptions } from './types';
+import { E_SSROptions } from './types';
 
-export const DEFAULTS: CoreOptions = {
+export const __PATHS__ = {
     pathRoot: 'views',
     templates: 'pages',
-    partials: 'partials',
+    partials: 'partials'
+};
+
+export const HCL_DEFAULTS: E_SSROptions = {
+    ...__PATHS__,
     templateInput: {},
     partialInput: {},
     watch: false,
+    discoverPaths: false,
+    intlCode: 'en',
 	debug: {
 		logMode: 'silent',
 		logStrategy: 'none',
         logFile: null
 	},
-    cacheExpiration: 0
-};
-
-export const __DEFAULTS = {
-	...DEFAULTS,
-    _publishDefault : 'dist',
-    outDefault: 'public',
-    static_config: {
-        pathRoot: 'views',
-        partials: 'partials',
-        templates: 'pages',
-        outPath: 'public',
-        loaderFile: 'loader.js',
-        cleanup: true
+    cache: {
+        ttl: 0 
     }
 };
 
+export const STATIC_DEFAULTS = {
+	...__PATHS__,
+    outPath: 'public',
+    loaderFile: 'loader.js',
+    cleanup: true
+};
+
+export const FULL_DEFAULTS = {
+    base_config: HCL_DEFAULTS,
+    static_config: STATIC_DEFAULTS,
+    fallbacks: {}
+};